+++ CHANGELOG.rst
@@ -1,4 +1,12 @@
 CHANGELOG
 =========
 
-# No releases done yet
+0.0.1 (2016-04-18)
+------------------
+
+* **Initial release**
+* Initial Position Model added
+* Added PositionPlugin Model
+* Added CMS Plugins (Container & Entry)
+* Plugin showcase in ReadMe added
+* Added empty templates